+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2022-08-15
+- Add integrity check for nonce too low
+
+- Add method for instantiating `tx-manager` with cleared database
+
 ## [0.5.0] - 2022-08-05
 - Remove sender address from configuration.
 
@@ -35,7 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-12-28
 - Initial release
 
-[Unreleased]: https://github.com/cartesi-corp/tx-manager/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/cartesi-corp/tx-manager/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/cartesi-corp/tx-manager/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/cartesi-corp/tx-manager/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/cartesi-corp/tx-manager/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/cartesi-corp/tx-manager/compare/v0.2.0...v0.3.0